12/15 가지의 버그/수정사항 수정, 2개의 타일에 걸쳐 있을 경우 왼쪽 타일을 우선해서 충돌체크 되는 로직을 수정

+++ Project-FW/MapBackground.cpp
@@ -85,7 +85,7 @@ void CMapBackground::Init()
 	m_nSize = m_MapBackground_List.size() ;
 
 	//m_fMapWidth = MapSize.x ;
-	m_fMapWidth = BackgroundSize.x * WinWidth ;
+	m_fMapWidth = (float)(BackgroundSize.x * WinWidth) ;
 }
 
 void CMapBackground::Scroll()
